Fix Kotlin type safety issues in Android plugin
- Fix nullable type handling in MorphProtocolPlugin.kt
  Use Elvis operator (?:) instead of default parameter values
- Fix Long vs Int comparison in MorphUdpClient.kt (0L instead of 0)
- Fix bitwise NOT operation in ObfuscationFunction.kt
  Convert to Int before inv() then back to Byte

Fixes compilation errors:
- Type mismatch: inferred type is String?/Int?/Long? but non-null was expected
- Operator '==' cannot be applied to 'Long' and 'Int'
- Unresolved reference inv() on Byte

+++ android/plugin/android/src/main/java/com/morphprotocol/client/core/functions/ObfuscationFunction.kt
@@ -150,7 +150,7 @@ class BitwiseNOT : ObfuscationFunction {
     override fun obfuscate(input: ByteArray, keyArray: ByteArray, initor: Any?): ByteArray {
         val output = ByteArray(input.size)
         for (i in input.indices) {
-            output[i] = input[i].inv()
+            output[i] = input[i].toInt().inv().toByte()
         }
         return output
     }
